// src/types/codes.ts
export enum CURRENCY_CODES {
    UAEDirham = 'AED',
    Afghani = 'AFN',
    Lek = 'ALL',
    ArmenianDram = 'AMD',
    NetherlandsAntilleanGuilder = 'ANG',
    Kwanza = 'AOA',
    ArgentinePeso = 'ARS',
    AustralianDollar = 'AUD',
    ArubanFlorin = 'AWG',
    AzerbaijanManat = 'AZN',
    ConvertibleMark = 'BAM',
    BarbadosDollar = 'BBD',
    Taka = 'BDT',
    BulgarianLev = 'BGN',
    BahrainiDinar = 'BHD',
    BurundiFranc = 'BIF',
    BermudianDollar = 'BMD',
    BruneiDollar = 'BND',
    Boliviano = 'BOB',
    Mvdol = 'BOV',
    BrazilianReal = 'BRL',
    BahamianDollar = 'BSD',
    Ngultrum = 'BTN',
    Pula = 'BWP',
    BelarusianRuble = 'BYN',
    BelizeDollar = 'BZD',
    CanadianDollar = 'CAD',
    CongoleseFranc = 'CDF',
    WIREuro = 'CHE',
    SwissFranc = 'CHF',
    WIRFranc = 'CHW',
    UnidaddeFomento = 'CLF',
    ChileanPeso = 'CLP',
    YuanRenminbi = 'CNY',
    ColombianPeso = 'COP',
    UnidaddeValorReal = 'COU',
    CostaRicanColon = 'CRC',
    PesoConvertible = 'CUC',
    CubanPeso = 'CUP',
    CaboVerdeEscudo = 'CVE',
    CzechKoruna = 'CZK',
    DjiboutiFranc = 'DJF',
    DanishKrone = 'DKK',
    DominicanPeso = 'DOP',
    AlgerianDinar = 'DZD',
    EgyptianPound = 'EGP',
    Nakfa = 'ERN',
    EthiopianBirr = 'ETB',
    Euro = 'EUR',
    FijiDollar = 'FJD',
    FalklandIslandsPound = 'FKP',
    PoundSterling = 'GBP',
    Lari = 'GEL',
    GhanaCedi = 'GHS',
    GibraltarPound = 'GIP',
    Dalasi = 'GMD',
    GuineanFranc = 'GNF',
    Quetzal = 'GTQ',
    GuyanaDollar = 'GYD',
    HongKongDollar = 'HKD',
    Lempira = 'HNL',
    Kuna = 'HRK',
    Gourde = 'HTG',
    Forint = 'HUF',
    Rupiah = 'IDR',
    NewIsraeliSheqel = 'ILS',
    IndianRupee = 'INR',
    IraqiDinar = 'IQD',
    IranianRial = 'IRR',
    IcelandKrona = 'ISK',
    JamaicanDollar = 'JMD',
    JordanianDinar = 'JOD',
    Yen = 'JPY',
    KenyanShilling = 'KES',
    Som = 'KGS',
    Riel = 'KHR',
    ComorianFranc = 'KMF',
    NorthKoreanWon = 'KPW',
    Won = 'KRW',
    KuwaitiDinar = 'KWD',
    CaymanIslandsDollar = 'KYD',
    Tenge = 'KZT',
    LaoKip = 'LAK',
    LebanesePound = 'LBP',
    SriLankaRupee = 'LKR',
    LiberianDollar = 'LRD',
    Loti = 'LSL',
    LibyanDinar = 'LYD',
    MoroccanDirham = 'MAD',
    MoldovanLeu = 'MDL',
    MalagasyAriary = 'MGA',
    Denar = 'MKD',
    Kyat = 'MMK',
    Tugrik = 'MNT',
    Pataca = 'MOP',
    Ouguiya = 'MRU',
    MauritiusRupee = 'MUR',
    Rufiyaa = 'MVR',
    MalawiKwacha = 'MWK',
    MexicanPeso = 'MXN',
    MexicanUnidaddeInversion = 'MXV',
    MalaysianRinggit = 'MYR',
    MozambiqueMetical = 'MZN',
    NamibiaDollar = 'NAD',
    Naira = 'NGN',
    CordobaOro = 'NIO',
    NorwegianKrone = 'NOK',
    NepaleseRupee = 'NPR',
    NewZealandDollar = 'NZD',
    RialOmani = 'OMR',
    Balboa = 'PAB',
    Sol = 'PEN',
    Kina = 'PGK',
    PhilippinePeso = 'PHP',
    PakistanRupee = 'PKR',
    Zloty = 'PLN',
    Guarani = 'PYG',
    QatariRial = 'QAR',
    RomanianLeu = 'RON',
    SerbianDinar = 'RSD',
    RussianRuble = 'RUB',
    RwandaFranc = 'RWF',
    SaudiRiyal = 'SAR',
    SolomonIslandsDollar = 'SBD',
    SeychellesRupee = 'SCR',
    SudanesePound = 'SDG',
    SwedishKrona = 'SEK',
    SingaporeDollar = 'SGD',
    SaintHelenaPound = 'SHP',
    Leone = 'SLL',
    SomaliShilling = 'SOS',
    SurinamDollar = 'SRD',
    SouthSudanesePound = 'SSP',
    Dobra = 'STN',
    ElSalvadorColon = 'SVC',
    SyrianPound = 'SYP',
    Lilangeni = 'SZL',
    Baht = 'THB',
    Somoni = 'TJS',
    TurkmenistanNewManat = 'TMT',
    TunisianDinar = 'TND',
    PaAnga = 'TOP',
    TurkishLira = 'TRY',
    TrinidadandTobagoDollar = 'TTD',
    NewTaiwanDollar = 'TWD',
    TanzanianShilling = 'TZS',
    Hryvnia = 'UAH',
    UgandaShilling = 'UGX',
    USDollar = 'USD',
    USDollarNextday = 'USN',
    UruguayPesoenUnidadesIndexadas = 'UYI',
    PesoUruguayo = 'UYU',
    UnidadPrevisional = 'UYW',
    UzbekistanSum = 'UZS',
    BolívarSoberano = 'VES',
    Dong = 'VND',
    Vatu = 'VUV',
    Tala = 'WST',
    CFAFrancBEAC = 'XAF',
    Silver = 'XAG',
    Gold = 'XAU',
    BondMarketsUnitEuropeanCompositeUnit = 'XBA',
    BondMarketsUnitEuropeanMonetaryUnit = 'XBB',
    BondMarketsUnitEuropeanUnitofAccount9 = 'XBC',
    BondMarketsUnitEuropeanUnitofAccount17 = 'XBD',
    EastCaribbeanDollar = 'XCD',
    SDR = 'XDR',
    CFAFrancBCEAO = 'XOF',
    Palladium = 'XPD',
    CFPFranc = 'XPF',
    Platinum = 'XPT',
    Sucre = 'XSU',
    TestingCode = 'XTS',
    ADBUnitofAccount = 'XUA',
    NoCurrencyInvolved = 'XXX',
    YemeniRial = 'YER',
    Rand = 'ZAR',
    ZambianKwacha = 'ZMW',
    ZimbabweDollar = 'ZWL'
}
// ...
export enum EAS_SCHEME_CODES {
    OFTP_ODETTE_File_Transfer_Protocol = 'AN',
    X_400_address_for_mail_text = 'AQ_',
    AS2_exchange_ = 'AS_',
    File_Transfer_Protocol = 'AU_',
    Electronic_mail_SMPT = 'EM',
    System_Information_et_Repertoire_des_Entreprise_et_des_Etablissements_SIRENE = '0002',
    Organisationsnummer = '0007',
    SIRET_CODE = '0009',
    LY_tunnus = '0037',
    Data_Universal_Numbering_System_DUNS_Number = '0060',
    EAN_Location_Code = '0088',
    DANISH_CHAMBER_OF_COMMERCE_Scheme_EDIRA_compliant = '0096',
    FTI_Ediforum_Italia_EDIRA_compliant = '0097',
    Vereniging_van_Kamers_van_Koophandel_en_Fabrieken_in_Nederland_Association_of_Chambers_of_Commerce_and_Industry_in_the_Netherlands_Scheme_EDIRA_compliant = '0106',
    Directorates_of_the_European_Commission = '0130',
    SIA_Object_Identifiers = '0135',
    SECETI_Object_Identifiers = '0142',
    Standard_Company_Code = '0147',
    Australian_Business_Number_ABN_Scheme = '0151',
    Teikoku_Company_Code = '0170',
    Numéro_d_identification_suisse_des_enterprises_IDE_Swiss_Unique_Business_Identification_Number_UIDB_ = '0183',
    DIGSTORG = '0184',
    Corporate_Number_of_The_Social_Security_and_Tax_Number_System = '0188',
    Dutch_Originators_Identification_Number = '0190',
    Centre_of_Registers_and_Information_Systems_of_the_Ministry_of_Justice = '0191',
    Enhetsregisteret_ved_Bronnoysundregisterne_ = '0192',
    UBL_BE_party_identifier = '0193',
    KOIOS_Open_Technical_Dictionary = '0194',
    Singapore_UEN_identifier = '0195',
    Kennitala_Iceland_legal_id_for_individuals_and_legal_entities = '0196',
    ERSTORG = '0198',
    Global_legal_entity_identifier_GLEIF = '0199',
    Legal_entity_code_Lithuania = '0200',
    Codice_Univoco_Unità_Organizzativa_iPA = '0201',
    Indirizzo_di_Posta_Elettronica_Certificata = '0202',
    eDelivery_Network_Participant_identifier = '0203',
    Leitweg_ID = '0204',
    CODDEST = '0205',
    Numero_d_entreprise_ondernemingsnummer_Unternehmensnummer = '0208',
    GS1_identification_keys = '0209',
    CODICE_FISCALE = '0210',
    PARTITA_IVA = '0211',
    Finnish_Organization_Identifier = '0212',
    Finnish_Organization_Value_Add_Tax_Identifier = '0213',
    Net_service_ID = '0215',
    OVTcode = '0216',
    The_Netherlands_Chamber_of_Commerce_and_Industry_establishment_number = '0217',
    Unified_registration_number_Latvia = '0218',
    The_registered_number_of_the_qualified_invoice_issuer = '0221',
    FRCTC_ELECTRONIC_ADDRESS = '0225',
    National_e_Invoicing_Framework = '0230',
    Danish_Ministry_of_the_Interior_and_Health = '9901',
    Hungary_VAT_number = '9910',
    Business_Registers_Network_ = '9913',
    Österreichische_Umsatzsteuer_Identifikationsnummer_ = '9914',
    Österreichisches_Verwaltungs_bzw_Organisationskennzeichen = '9915',
    SOCIETY_FOR_WORLDWIDE_INTERBANK_FINANCIAL_TELECOMMUNICATION_SWIFT = '9918',
    Kennziffer_des_Unternehmensregisters_ = '9919',
    Agencia_Española_de_Administración_Tributaria_ = '9920',
    Andorra_VAT_number = '9922',
    Albania_VAT_number = '9923',
    Bosnia_and_Herzegovina_VAT_number = '9924',
    Belgium_VAT_number = '9925',
    Bulgaria_VAT_number = '9926',
    Switzerland_VAT_number = '9927',
    Cyprus_VAT_number = '9928',
    Czech_Republic_VAT_number = '9929',
    Germany_VAT_number = '9930',
    Estonia_VAT_number = '9931',
    United_Kingdom_VAT_number = '9932',
    Greece_VAT_number = '9933',
    Croatia_VAT_number = '9934',
    Ireland_VAT_number = '9935',
    Liechtenstein_VAT_number = '9936',
    Lithuania_VAT_number = '9937',
    Luxemburg_VAT_number = '9938',
    Latvia_VAT_number = '9939',
    Monaco_VAT_number = '9940',
    Montenegro_VAT_number = '9941',
    Macedonia_the_former_Yugoslav_Republic_of_VAT_number = '9942',
    Malta_VAT_number = '9943',
    Netherlands_VAT_number = '9944',
    Poland_VAT_number = '9945',
    Portugal_VAT_number = '9946',
    Romania_VAT_number = '9947',
    Serbia_VAT_number = '9948',
    Slovenia_VAT_number = '9949',
    Slovakia_VAT_number = '9950',
    San_Marino_VAT_number = '9951',
    Turkey_VAT_number = '9952',
    Holy_See_Vatican_City_State_VAT_number = '9953',
    French_VAT_number = '9957',
    Employer_Identification_Number_EIN_USA = '9959'
}

export enum TAX_CODES {
    Value_Added_Tax_VAT = 'VA',
    Fiscal_Number = 'FC'
}

export enum TAX_TYPE_CODE {
    VALUE_ADDED_TAX_VAT = 'VAT'
}
// ...
export enum ALLOWANCE_REASONS_CODES {
    Bonus_for_works_ahead_of_schedule = '41',
    Other_bonus = '42',
    Manufacturers_consumer_discount = '60',
    Due_to_military_status = '62',
    Due_to_work_accident = '63',
    Special_agreement = '64',
    Production_error_discount = '65',
    New_outlet_discount = '66',
    Sample_discount = '67',
    End_of_range_discount = '68',
    Incoterm_discount = '70',
    Point_of_sales_threshold_allowance = '71',
    Material_surcharge_deduction = '88',
    Discount = '95',
    Special_rebate = '100',
    Fixed_long_term = '102',
    Temporary = '103',
    Standard = '104',
    Yearly_turnover = '105'
}
// ...
export enum DOCUMENT_TYPE_CODES {
    REQUEST_FOR_PAYMENT = '71',
    DEBIT_NOTE_RELATED_TO_GOODS_OR_SERVICES = '80',
    CREDIT_NOTE_RELATED_TO_GOODS_OR_SERVICES = '81',
    METERED_SERVICES_INVOICE = '82',
    CREDIT_NOTE_RELATED_TO_FINANCIAL_ADJUSTMENTS = '83',
    DEBIT_NOTE_RELATED_TO_FINANCIAL_ADJUSTMENTS = '84',
    TAX_NOTIFICATION = '102',
    INVOICING_DATA_SHEET = '130',
    DIRECT_PAYMENT_VALUATION = '202',
    PROVISIONAL_PAYMENT_VALUATION = '203',
    PAYMENT_VALUATION = '204',
    INTERIM_APPLICATION_FOR_PAYMENT = '211',
    FINAL_PAYMENT_REQUEST_BASED_ON_COMPLETION_OF_WORK = '218',
    PAYMENT_REQUEST_FOR_COMPLETED_UNITS = '219',
    SELF_BILLED_CREDIT_NOTE = '261',
    CONSOLIDATED_CREDIT_NOTE___GOODS_AND_SERVICES = '262',
    PRICE_VARIATION_INVOICE = '295',
    CREDIT_NOTE_FOR_PRICE_VARIATION = '296',
    DELCREDERE_CREDIT_NOTE = '308',
    PROFORMA_INVOICE = '325',
    PARTIAL_INVOICE = '326',
    COMMERCIAL_INVOICE_WHICH_INCLUDES_A_PACKING_LIST = '331',
    COMMERCIAL_INVOICE = '380',
    CREDIT_NOTE = '381',
    COMMISSION_NOTE = '382',
    DEBIT_NOTE = '383',
    CORRECTED_INVOICE = '384',
    CONSOLIDATED_INVOICE = '385',
    PREPAYMENT_INVOICE = '386',
    HIRE_INVOICE = '387',
    TAX_INVOICE = '388',
    SELF_BILLED_INVOICE = '389',
    DELCREDERE_INVOICE = '390',
    FACTORED_INVOICE = '393',
    LEASE_INVOICE = '394',
    CONSIGNMENT_INVOICE = '395',
    FACTORED_CREDIT_NOTE = '396',
    OPTICAL_CHARACTER_READING_OCR_PAYMENT_CREDIT_NOTE = '420',
    DEBIT_ADVICE = '456',
    REVERSAL_OF_DEBIT = '457',
    REVERSAL_OF_CREDIT = '458',
    SELF_BILLED_DEBIT_NOTE = '527',
    FORWARDERS_CREDIT_NOTE = '532',
    FORWARDERS_INVOICE_DISCREPANCY_REPORT = '553',
    INSURERS_INVOICE = '575',
    FORWARDERS_INVOICE = '623',
    PORT_CHARGES_DOCUMENTS = '633',
    INVOICE_INFORMATION_FOR_ACCOUNTING_PURPOSES = '751',
    FREIGHT_INVOICE = '780',
    CLAIM_NOTIFICATION = '817',
    CONSULAR_INVOICE = '870',
    PARTIAL_CONSTRUCTION_INVOICE = '875',
    PARTIAL_FINAL_CONSTRUCTION_INVOICE = '876',
    FINAL_CONSTRUCTION_INVOICE = '877',
    CUSTOMS_INVOICE = '935'
}